Full update.
Add multiple functions and optimize existing ones.

+++ open_ahrs/src/common.rs
@@ -94,9 +94,10 @@ pub fn generate_random_attitudes(number: u8) -> Result<Matrix, LinalgError> {
     let mut small_rng = SmallRng::seed_from_u64(0u64);
 
     for n in 0..number {
-        let u1: f32 = small_rng.next_u64() / u64::MAX;  // Generates a random number between 0 and 1.
-        let u2: f32 = small_rng.next_u64() / u64::MAX;  // Generates a random number between 0 and 1.
-        let u3: f32 = small_rng.next_u64() / u64::MAX;  // Generates a random number between 0 and 1.
+        // Warning  here.
+        let u1: f32 = (small_rng.next_u64() / u64::MAX) as f32; // Generates a random number between 0 and 1.
+        let u2: f32 = (small_rng.next_u64() / u64::MAX) as f32; // Generates a random number between 0 and 1.
+        let u3: f32 = (small_rng.next_u64() / u64::MAX) as f32; // Generates a random number between 0 and 1.
 
         let s1 = sqrtf(1.0_f32 - u1);
         let s2 = sqrtf(u1);
